Changed signature of DistanceFunction and added L1 distance.

+++ src/main/kotlin/ch/unibas/dmi/dbis/cottontail/knn/metrics/DistanceFunction.kt
@@ -4,7 +4,7 @@ interface DistanceFunction {
     /**
      *
      */
-    operator fun invoke(a: FloatArray, b: FloatArray): Float
+    operator fun invoke(a: FloatArray, b: FloatArray): Double
 
     /**
      *
